deps: Upgrade firebase_core, firebase_messaging to latest
This commit is the result of the following commands:
  flutter pub upgrade --major-versions firebase_messaging firebase_core
  tools/upgrade pod

Changelogs:
  https://pub.dev/packages/firebase_core/changelog#410
  https://pub.dev/packages/firebase_messaging/changelog#1601

Notable changes include bump to Firebase Android BoM (34.0.0 to 34.1.0)
and Firebase iOS SDK (12.0.0 to 12.2.0), changelog for those are at:
  https://firebase.google.com/support/release-notes/android
  https://firebase.google.com/support/release-notes/ios

For Android SDK, no changes in the FCM component (the only one we use).

For iOS SDK, one runtime crash fix for apps built with
(not yet released) Xcode 26:
  https://github.com/firebase/firebase-ios-sdk/issues/15159

which doesn't affect us, as probably no one has tried to build
with Xcode 26 Beta yet.